VC warning fixes
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@3465 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/console.cpp
@@ -105,7 +105,7 @@ int rendercommand(int x, int y, int w)
     int width, height;
     text_bounds(s, width, height, w);
     y -= height - FONTH;
-    draw_text(s, x, y, 0xFF, 0xFF, 0xFF, 0xFF, cmdline.pos>=0 ? cmdline.pos+2 : strlen(s), w);
+    draw_text(s, x, y, 0xFF, 0xFF, 0xFF, 0xFF, cmdline.pos>=0 ? cmdline.pos+2 : (int)strlen(s), w);
     return height;
 }
